backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-storages[s3]==1.14.4
psycopg2-binary==2.9.9
pgvector==0.3.2
celery==5.4.0
redis==5.0.8
anthropic==0.34.0
openai==1.45.0
pymupdf==1.24.10
python-docx==1.1.2
python-decouple==3.8
Pillow==10.4.0
gunicorn==23.0.0
